+++ docs/src/components/TextInput/TextInput.js
@@ -54,5 +54,13 @@ export default () => (
       label="With suffix read only input"
       readOnly
     />
+    <TextInput
+      name="focusable"
+      label="Focusable Input"
+      value="this is a focusable input"
+      onFocus={(name, value) =>
+        console.log(`${name} focused with value: ${value}`)
+      }
+    />
   </Component>
 );